Leapmvp/backend/app/services/ml_service.py
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from datetime import datetime, timedelta
from typing import List, Dict, Optional
import logging

class MLService:
    def __init__(self):
        try:
            self.priority_model = RandomForestClassifier()
            self.schedule_model = None
            self._initialize_models()
        except Exception as e:
            logging.error(f"Failed to initialize ML models: {str(e)}")
            raise

    def _initialize_models(self) -> None:
        # Initialize and load pre-trained models if available
        pass

    async def prioritize_tasks(
        self, 
        tasks: List[Dict], 
        user_preferences: Dict
    ) -> List[Dict]:
        """
        Prioritize tasks based on various factors
        """
        try:
            features = self._extract_features(tasks)
            priorities = self._calculate_priorities(features, user_preferences)
            
            prioritized_tasks = []
            for task, priority in zip(tasks, priorities):
                task["priority"] = priority
                prioritized_tasks.append(task)
            
            return sorted(prioritized_tasks, key=lambda x: x["priority"], reverse=True)
        except Exception as e:
            logging.error(f"Error prioritizing tasks: {str(e)}")
            return tasks  # Return original tasks if prioritization fails

    async def suggest_time_slot(
        self, 
        task: Dict, 
        existing_schedule: List[Dict]
    ) -> Optional[datetime]:
        """
        Suggest optimal time slot for a task
        """
        try:
            available_slots = self._find_available_slots(existing_schedule)
            return self._select_optimal_slot(task, available_slots)
        except Exception as e:
            logging.error(f"Error suggesting time slot: {str(e)}")
            return None

    def _extract_features(self, tasks: List[Dict]) -> np.ndarray:
        features = []
        for task in tasks:
            feature_vector = [
                task.get("estimated_duration", 30),
                task.get("due_date", datetime.now()).timestamp(),
                len(task.get("title", "")),
                task.get("priority", 2)
            ]
            features.append(feature_vector)
        return np.array(features)

    def _calculate_priorities(
        self, 
        features: np.ndarray, 
        user_preferences: Dict
    ) -> List[int]:
        # Apply user preferences and calculate priority scores
        scores = np.mean(features, axis=1)  # Simplified scoring
        return [int(score) for score in scores]

    def _find_available_slots(
        self, 
        schedule: List[Dict]
    ) -> List[datetime]:
        # Find free time slots in the schedule
        current_time = datetime.now()
        available_slots = []
        
        # Add logic to find available time slots
        return available_slots

    def _select_optimal_slot(
        self, 
        task: Dict, 
        available_slots: List[datetime]
    ) -> Optional[datetime]:
        if not available_slots:
            return None
        return min(available_slots, key=lambda x: abs(x - datetime.now())) 