Leapmvp/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
torch>=2.0.0
torchvision>=0.15.0
pytest>=7.0.0
numpy>=1.21.0
python-dateutil>=2.8.2
spacy>=3.0.0
scikit-learn>=1.0.0
transformers>=4.0.0 